implemented template method pattern for window
